+++ Workers/Workers.Web/Areas/Admin/Controllers/EmployeeController.cs
@@ -2,9 +2,6 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Workers.Web.Infrastructure.Context;
 using Workers.Web.Infrastructure.Models;
