+++ Workers/Workers.Web/Areas/Admin/Views/Project/Delete.cshtml
