+++ Workers/Workers.Web/Areas/Admin/Views/StatusTask/Delete.cshtml
