+++ Workers/Workers.Web/Controllers/StatusTaskController.cs
@@ -27,20 +27,6 @@ namespace Workers.Web.Controllers
             return View(statusTasks);
         }
 
-        [HttpGet("create")]
-        public IActionResult Create()
-        {
-            return View();
-        }
-
-        [HttpPost("create")]
-        public async Task<IActionResult> Create(Status statusTask)
-        {
-            await _db.AddAsync(statusTask);
-            await _db.SaveChangesAsync();
-            return LocalRedirect("~/status-task/all");
-        }
-
         [HttpGet("id")]
         public async Task<IActionResult> GetDetails(int id)
         {
@@ -51,57 +37,5 @@ namespace Workers.Web.Controllers
             }
             return View(statusTasks);
         }
-
-        [HttpGet("edit/{id}")]
-        public async Task<IActionResult> Edit(int id)
-        {
-            var statusTask = await _db.Statuses.SingleOrDefaultAsync(x =>x.Id == id);
-            if(statusTask == null)
-            {
-                return LocalRedirect("~/status-task/all");
-            }
-            return View(statusTask);
-        }
-
-        [HttpPost("edit")]
-        public async Task<IActionResult> EditStatusTask(Status statusTask)
-        {
-            var statusTaskFromDb = await _db.Statuses.SingleOrDefaultAsync(x => x.Id == statusTask.Id);
-            if (statusTaskFromDb == null)
-            {
-                return LocalRedirect("~/status-task/all");
-            }
-            statusTaskFromDb.StatusName = statusTask.StatusName;
-            _db.Update(statusTaskFromDb);
-            await _db.SaveChangesAsync();
-            return LocalRedirect("~/status-task/all");
-        }
-
-        [HttpGet("delete/{id}")]
-        public async Task<IActionResult> Delete (int id)
-        {
-            var statusTask = await _db.Statuses.SingleOrDefaultAsync(x => x.Id == id);
-            if (statusTask == null)
-            {
-                return LocalRedirect("~/status-task/all");
-            }
-            return View(statusTask);
-        }
-
-        [HttpPost("delete")]
-        public async Task<IActionResult> DeleteTask(Status statusTask)
-        {
-            var statusTaskFromDb = await _db.Statuses.SingleOrDefaultAsync(x => x.Id == statusTask.Id);
-            if (statusTaskFromDb == null)
-            {
-                return LocalRedirect("~/status-task/all");
-            }
-            _db.Remove(statusTaskFromDb);
-            await _db.SaveChangesAsync();
-            return LocalRedirect("~/status-task/all");
-        }
-
-
-
     }
 }
