+++ Workers/Workers.Web/Controllers/TaskController.cs
@@ -16,7 +16,7 @@ namespace Workers.Web.Controllers
         {
             _db = db;
         }
-        [Authorize(Policy = "ForEmail")]
+
         [HttpGet("all")]
         public async Task<IActionResult> GetAll()
         {
@@ -30,86 +30,5 @@ namespace Workers.Web.Controllers
             var task = await _db.Tasks.Include(x => x.TaskStatuses).ThenInclude(x => x.StatusTask).SingleOrDefaultAsync(x =>x.Id == id);
             return View(task);
         }
-
-        [HttpGet("create")]
-        public async Task<IActionResult> Create()
-        {
-            ViewBag.Projects = await _db.Projects.ToListAsync();
-            ViewBag.StatusTasks = await _db.Statuses.ToListAsync();
-            return View();
-        }
-
-        [HttpPost("create")]
-        public async Task<IActionResult> Create(Infrastructure.Models.Task task)
-        {
-            await _db.Tasks.AddAsync(task);
-            await _db.SaveChangesAsync();
-            return LocalRedirect("~/task/all");
-        }
-
-        [HttpGet("edit/{id}")]
-        public async Task<IActionResult> Edit(int id)
-        {
-            var task = await _db.Tasks.Include(x => x.Project).Include(x => x.TaskStatuses).SingleOrDefaultAsync(x => x.Id == id);
-            if(task == null)
-            {
-                return LocalRedirect("~/task/all");
-            }
-
-            var projects = await _db.Projects.ToArrayAsync();
-
-            ViewBag.Projects = projects;
-
-            return View(task);
-        }
-
-        [HttpPost("edit")]
-        public async Task<IActionResult> EditTask(Infrastructure.Models.Task task)
-        {
-            var taskFromDb = await _db.Tasks.Include(x => x.Project).Include(x => x.TaskStatuses).SingleOrDefaultAsync(x => x.Id == task.Id);
-            if (task == null)
-            {
-                return LocalRedirect("~/task/all");
-            }
-
-            taskFromDb.DeadLine = task.DeadLine;
-            taskFromDb.Name = task.Name;
-            taskFromDb.PercentOfCompleted = task.PercentOfCompleted;
-            taskFromDb.Description = task.Description;
-            taskFromDb.ProjectId = task.ProjectId;
-
-            _db.Update(taskFromDb);
-            await _db.SaveChangesAsync();
-            return LocalRedirect("~/task/all");
-        }
-
-        [HttpGet("delete/{id}")]
-        public async Task<IActionResult> Delete (int id)
-        {
-            var task = await _db.Tasks.Include(x => x.Project).Include(x =>x.TaskStatuses).SingleOrDefaultAsync(x => x.Id == id);
-            if (task == null)
-            {
-                return LocalRedirect("~/task/all");
-            }
-            return View(task);
-        }
-
-        [HttpPost("delete")]
-        public async Task<IActionResult> DeleteTask(Infrastructure.Models.Task task)
-        {
-            var taskFromDb = await _db.Tasks
-                .Include(x => x.Project)
-                .Include(x => x.TaskStatuses)
-                .SingleOrDefaultAsync(x => x.Id == task.Id);
-            if (task == null)
-            {
-                return LocalRedirect("~/task/all");
-            }
-            _db.Remove(taskFromDb);
-            await _db.SaveChangesAsync();
-            return LocalRedirect("~/task/all");
-        }
-
-
     }
 }
