+++ Workers/Workers.Web/Views/Employee/GetAll.cshtml
@@ -7,10 +7,6 @@
 
 <h1>GetAll</h1>
 
-@*@if (User.IsInRole("Admin"))
-    {
-
-    }*@
 <p>
     <a asp-area="Admin" asp-controller="Employee" asp-action="Create">Create New</a>
 </p>
