+++ Workers/Workers.Web/Views/ProjectEmployee/GetAll.cshtml
@@ -2,12 +2,13 @@
 
 @{
     ViewData["Title"] = "GetAll";
+    Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
 <h1>GetAll</h1>
 
 <p>
-    <a asp-action="Create">Create New</a>
+    <a asp-area="Admin" asp-controller="ProjectEmployee" asp-action="Create">Create New</a>
 </p>
 @if (Model.Any())
 {
@@ -45,11 +46,14 @@
                     <td>
                         @Html.ActionLink(item.Position.Name, "GetDetails", "Position", new { id = item.PositionId })
                     </td>
-                    <td>
-                        @Html.ActionLink("Edit", "Edit", new { id = item.Id }) |
-                        @Html.ActionLink("Details", "GetDetails", "ProjectEmployee", new { id = item.Id }) |
-                        @Html.ActionLink("Delete", "Delete", new { id = item.Id })
-                    </td>
+                    @if (User.IsInRole("Admin"))
+                    {
+                        <td>
+                            @Html.ActionLink("Edit", "Edit", "ProjectEmployee", new { id = item.Id, Area = "Admin" }) |
+                            @Html.ActionLink("Details", "GetDetails", "ProjectEmployee", new { id = item.Id, Area = "Admin" }) |
+                            @Html.ActionLink("Delete", "Delete", "ProjectEmployee", new { id = item.Id, Area = "Admin" })
+                        </td>
+                    }
                 </tr>
             }
         </tbody>
