+++ Workers/Workers.Web/Views/Shared/_Layout.cshtml
@@ -45,13 +45,13 @@
                             <li class="nav-item">
                                 <a class="nav-link text-dark" asp-controller="TaskStatus" asp-action="GetAll">TaskStatus</a>
                             </li>
-                            <li class="nav-item">
-                                <a class="nav-link text-dark" asp-controller="EmployeesRole" asp-action="GetAll">EmployeesRole</a>
-                            </li>
                             @if (User.IsInRole("Admin"))
                             {
                                 <li class="nav-item">
                                     <a class="nav-link text-dark" asp-area="Admin" asp-controller="Admin" asp-action="GetAll">Admin</a>
+                                </li>
+                                <li class="nav-item">
+                                    <a class="nav-link text-dark" asp-controller="EmployeesRole" asp-action="GetAll">EmployeesRole</a>
                                 </li>
                             }
                         }
