+++ Workers/Workers.Web/Views/StatusProject/GetAll.cshtml
@@ -2,12 +2,13 @@
 
 @{
     ViewData["Title"] = "GetAll";
+    Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
 <h1>GetAll</h1>
 
 <p>
-    <a asp-action="Create">Create New</a>
+    <a asp-area="Admin" asp-controller="StatusProject" asp-action="Create">Create New</a>
 </p>
 @if (Model.Any())
 {
@@ -28,16 +29,16 @@
                         @Html.DisplayFor(modelItem => item.Status)
                     </td>
                     <td>
-                        <a asp-action="Edit" asp-route-id="@item.Id">Edit</a> |
-                        <a asp-action="Delete" asp-route-id="@item.Id">Delete</a>
+                        @Html.ActionLink("Edit", "Edit", "StatusProject", new { id = item.Id, area = "Admin" }) |
+                        @Html.ActionLink("Delete", "Delete", "StatusProject", new { id = item.Id, area = "Admin" }) 
                     </td>
                 </tr>
             }
         </tbody>
     </table>
 }
-            else
-            {
-                <p>No statuses!</p>
-            }
+else
+{
+    <h2>No statuses!</h2>
+}
