+++ Workers/Workers.Web/Views/Task/GetAll.cshtml
@@ -1,13 +1,14 @@
 ﻿@model IEnumerable<Workers.Web.Infrastructure.Models.Task>
 
 @{
+    Layout = "~/Views/Shared/_Layout.cshtml";
     ViewData["Title"] = "GetAll";
 }
 
 <h1>GetAll</h1>
 
 <p>
-    <a asp-action="Create">Create New</a>
+    <a asp-area="Admin" asp-controller="Task" asp-action="Create">Create New</a>
 </p>
 @if (Model.Any())
 {
@@ -52,9 +53,9 @@
                         @Html.DisplayFor(modelItem => item.Project.Name)
                     </td>
                     <td>
-                        <a asp-action="Edit" asp-route-id="@item.Id">Edit</a> |
-                        <a asp-action="GetDetails" asp-route-id="@item.Id">Details</a> |
-                        <a asp-action="Delete" asp-route-id="@item.Id">Delete</a>
+                        @Html.ActionLink("Edit", "Edit", "Task", new { id = item.Id, area = "Admin" }) |
+                        @Html.ActionLink("Details", "GetDetails", "Task", new { id = item.Id, area = "Admin" }) |
+                        @Html.ActionLink("Delete", "Delete", "Task", new { id = item.Id, area = "Admin" })
                     </td>
                 </tr>
             }
