+++ Workers/Workers.Web/Views/TaskStatus/GetAll.cshtml
@@ -2,12 +2,13 @@
 
 @{
     ViewData["Title"] = "GetAll";
+    Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
 <h1>GetAll</h1>
 
 <p>
-    <a asp-action="Create">Create New</a>
+    <a asp-area="Admin" asp-controller="TaskStatus" asp-action="Create">Create New</a>
 </p>
 @if (Model.Any())
 {
@@ -32,30 +33,33 @@
         <tbody>
             @foreach (var item in Model)
             {
-                <tr>
+            <tr>
+                <td>
+                    @Html.DisplayFor(modelItem => item.StatusTask.StatusName)
+                </td>
+                <td>
+                    @Html.DisplayFor(modelItem => item.Task.Name)
+                </td>
+                <td>
+                    @Html.DisplayFor(modelItem => item.DateEdit)
+                </td>
+                <td>
+                    @Html.DisplayFor(modelItem => item.Employee.FullName)
+                </td>
+                @if (User.IsInRole("Admin"))
+                {
                     <td>
-                        @Html.DisplayFor(modelItem => item.StatusTask.StatusName)
+                        @Html.ActionLink("Edit", "Edit", "TaskStatus", new { id = item.Id, area = "Admin" }) |
+                        @Html.ActionLink("Delete", "Delete", "TaskStatus", new { id = item.Id, area = "Admin" })
                     </td>
-                    <td>
-                        @Html.DisplayFor(modelItem => item.Task.Name)
-                    </td>
-                    <td>
-                        @Html.DisplayFor(modelItem => item.DateEdit)
-                    </td>
-                    <td>
-                        @Html.DisplayFor(modelItem => item.Employee.FullName)
-                    </td>
-                    <td>
-                        <a asp-action="Edit" asp-route-id="@item.Id">Edit</a> |
-                        <a asp-action="GetDetails" asp-route-id="@item.Id">Details</a> |
-                        <a asp-action="Delete" asp-route-id="@item.Id">Delete</a>
-                    </td>
-                </tr>
+                }
+                @Html.ActionLink("GetDetails", "GetDetails", "TaskStatus", new { id = item.Id })
+            </tr>
             }
         </tbody>
     </table>
 }
 else
 {
-    <p>No task status!</p>
+    <h2>No task status!</h2>
 }
